Move AzureFhirServerCredentialManager class back to fhir module
