+++ src/lib/phdi-building-blocks/phdi_building_blocks/fhir.py
@@ -4,13 +4,72 @@ import logging
 import polling
 import requests
 
+from datetime import datetime, timezone
 from requests.adapters import HTTPAdapter
 from typing import Union, Iterator, Tuple, TextIO
 from urllib3 import Retry
 
+from azure.core.credentials import AccessToken
 from azure.identity import DefaultAzureCredential
 from azure.storage.blob import download_blob_from_url
-from phdi_building_blocks.classes.creds_manager import AzureFhirServerCredentialManager
+
+
+class AzureFhirServerCredentialManager:
+    """
+    A class that manages handling Azure credentials for access to the FHIR server.
+    """
+
+    def __init__(self, fhir_url: str):
+        """
+        Constructor to create a credential manager.
+
+        :param fhir_url: The url of the FHIR server to-access
+        """
+        self.access_token = None
+        self.fhir_url = fhir_url
+
+    def get_fhir_url(self) -> str:
+        """
+        Getter to retrieve the FHIR URL.
+        :return: The FHIR url
+        """
+        return self.fhir_url
+
+    def get_access_token(self, token_reuse_tolerance: float = 10.0) -> AccessToken:
+        """
+        Obtain an access token for the FHIR server the manager is pointed at.
+        If the token is already set for this object and is not about to expire
+        (within token_reuse_tolerance parameter), then return the existing token.
+        Otherwise, request a new one.
+
+        :param token_reuse_tolerance: Number of seconds before expiration; it
+        is okay to reuse the currently assigned token
+        """
+        if not self._need_new_token(token_reuse_tolerance):
+            return self.access_token
+
+        # Obtain a new token if ours is going to expire soon
+        creds = DefaultAzureCredential()
+        scope = f"{self.fhir_url}/.default"
+        self.access_token = creds.get_token(scope)
+        return self.access_token
+
+    def _need_new_token(self, token_reuse_tolerance: float = 10.0) -> bool:
+        """
+        Determine whether the token already stored for this object can be reused,
+        or if it needs to be re-requested.
+
+        :param token_reuse_tolerance: Number of seconds before expiration
+        :return: Whether we need a new token (True means we do)
+        """
+        try:
+            current_time_utc = datetime.now(timezone.utc).timestamp()
+            return (
+                self.access_token.expires_on - token_reuse_tolerance
+            ) < current_time_utc
+        except AttributeError:
+            # access_token not set
+            return True
 
 
 def generate_filename(blob_name: str, message_index: int) -> str:
@@ -25,16 +84,6 @@ def generate_filename(blob_name: str, message_index: int) -> str:
     filename = blob_name.split("/")[-1]
     root_name, _ = filename.rsplit(".", 1)
     return f"{root_name}-{message_index}"
-
-
-def get_fhirserver_cred_manager(fhir_url: str) -> AzureFhirServerCredentialManager:
-    """
-    Get an instance of the Azure FHIR Server credential manager configured
-    for a given FHIR server url.
-
-    :param fhir_url: The url of the FHIR server to access
-    """
-    return AzureFhirServerCredentialManager(fhir_url)
 
 
 def upload_bundle_to_fhir_server(
