+++ src/lib/phdi-building-blocks/tests/test_fhir.py
@@ -9,7 +9,7 @@ from unittest import mock
 from azure.identity import DefaultAzureCredential
 
 from phdi_building_blocks.fhir import (
-    get_fhirserver_cred_manager,
+    AzureFhirServerCredentialManager,
     upload_bundle_to_fhir_server,
     export_from_fhir_server,
     _compose_export_url,
@@ -56,7 +56,7 @@ def test_get_access_token_reuse(mock_get_token):
 
     mock_get_token.return_value = mock_access_token
 
-    fhirserver_cred_manager = get_fhirserver_cred_manager("https://fhir-url")
+    fhirserver_cred_manager = AzureFhirServerCredentialManager("https://fhir-url")
     token1 = fhirserver_cred_manager.get_access_token()
 
     # Use the default token reuse tolerance, which is less than
@@ -75,7 +75,7 @@ def test_get_access_token_refresh(mock_get_token):
 
     mock_get_token.return_value = mock_access_token
 
-    fhirserver_cred_manager = get_fhirserver_cred_manager("https://fhir-url")
+    fhirserver_cred_manager = AzureFhirServerCredentialManager("https://fhir-url")
     token1 = fhirserver_cred_manager.get_access_token()
 
     # This time, use a very high token reuse tolerance to
